--- tests/menu.py
@@ -1,15 +1,0 @@
-import unittest
-#from tp_edd_2020.tests.menu import Menu
-
-class TestStringMethods(unittest.TestCase):
-
-    def test_upper(self):
-        self.assertEqual('foo'.upper(), 'FOO')
-
-    def test_isupper(self):
-        self.assertTrue('FOO'.isupper())
-        self.assertFalse('Foo'.isupper())
-
-
-if __name__ == '__main__':
-    unittest.main()
